+++ tests/YTWatchmarker2YTTakeout_test.py
@@ -4,13 +4,13 @@ import sys
 import os
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-import YTWatchmarker2Takeout
+import YTWatchmarker2YTTakeout
 
 def test_convestion():
     takeout = open('./tests/takeout.json')
     expected_list = json.load(takeout)
 
-    watchmaker = YTWatchmarker2Takeout.convert('./tests/YTWatchmaker.database')
+    watchmaker = YTWatchmarker2YTTakeout.convert('./tests/YTWatchmaker.database')
 
     for expected in expected_list:
         el = next((x for x in watchmaker if x.titleUrl == expected.get("titleUrl")), None)
